Update plugin for Swift 3.1 code generation
This required some extensive changes to template processing.
Alongside the Swift 3.0->3.1 update, the Swift Protocol Buffer
project changed the representation of "proto2" format messages
so that message fields no longer have fields in the generated
structures. This breaks the introspection that Stencil uses
during template rendering, so to fix this, we introduce an
intermediate structure.