Change the `ServerSession`s such that errors thrown during operation are returned to the client.
If you throw an error of type `ServerStatus`, that is returned to the client. For other errors, `ServerStatus.processingError` is returned.

This change also extends the interface of the `ServerSession`s to accept a custom status and a completion handler when closing the stream.
It also makes closing the stream mandatory for `ServerSessionBidirectionalStreaming`, `ServerSessionClientStreaming`, and `ServerSessionServerStreaming` (in fact it was already mandatory for the first two).

Also adds tests to verify this behavior.