Fix the GRPCTests again without using a sleep call:
We need to return status OK here, as it seems like the server might never send out the last few messages once it has been asked to send a non-OK status. Alternatively, we could send a non-OK status here, but then we would need to sleep for a few milliseconds before sending the non-OK status.