Fix a few problems found running "pod spec lint".
This includes updating the minimum supported iOS and OSX versions
to match gRPC and a fix for a compiler warning found by the linter.

The compiler warning was due to careless conversion of grpc event tags
from void * to various integer types that were 64 bit but inappropriate
for 32 bit architectures. The change now directly exposes void * tags
from the shim layer and converts them from UnsafeMutableRawPointers
to Int (and back) in SwiftGRPC/Core.