Spin up dedicated queues for some operations.
Without this, libDispatch on Linux waits for up to 1 second to spin up new threads for handlers when needed. This tremendously increases latency (by more than 1 second) on incoming requests if there have been no requests for a few seconds. This is because no-longer-used threads are discarded after a few seconds, so the 1-second wait time mentioned earlier applies.

Given that we need to spawn two fresh threads for each incoming request, anyway (one of the event-processing CompletionQueue and one for the actual code that responds to the request), we avoid these wait times by creating fresh dispatch queues (which each spin up a dedicated thread immediately instead of waiting) for each request handler.

This reduces the initial latency for "cold" requests from ~1.5s down to ~60ms. The remaining latency is probably the time needed by libDispatch to spin up new pthreads for the fresh queues. Still more than I'd like, but the need for fresh pthreads will only cease once we switch to a non-blocking event handling system like SwiftNIO.