Add make target generating an xcodeproj for Carthage.
An xcodeproj needs to be present in the repo for Carthage to work.
Only 4 frameworks are enabled for Carthage builds:

- BoringSSL
- CgRPC
- SwiftGRPC
- SwiftProtobuf

The other targets are removed from SwiftGRPC-Carthage.xcodeproj.

Moreover, a prebuild script is added to the build phases of the
cartage xcodeproj file in order to download the dependencies required
by Package.swift.

The Carthage-compatible xcodeproj can be generated with the following:

    $ make project-carthage