Rename SwiftGRPC to GRPC (#472)
* Rename SwiftGRPCNIO -> GRPC

* Update imports

* Update other references to SwiftGRPC

* Rename some test files

* Update Makefile

* rename NaturalLanguageNIO -> NaturalLanguage

* Update import ordering for Echo

* Update readme

* Increase the timeout for client timeout tests.

testBidirectionalStreamingTimeoutAfterSending has started failing on
Linux, the most likely explanation for this is the change which
schedules the timeout task on channelActive(context:) as opposed to when
the call to create a new stream is made. Since the test starts the call,
sends a message (which registers a callback to send the message once the
stream becomes active), waits and then sends end, it is possible for the
timeout to be registered once the wait has begun, thus end being sent
before the timeout has elapsed.

On one hand, this change makes the timeout closer to the point where
the server is sent the timout. On the other hand, a call could exceed the
timeout if it takes a long time for the stream to become active.