Add Server.Configuration (#490)
Motivation:

We added Configuration to the ClientConnection to make it simpler to
create new connections. We should have the equivalent for Server.

Modification:

Added a Configuration for Server similar to
ClientConnection.Configuration. This also made it possible to add
support for NIOTS for the server.

Result:

Configuring a server is easier and they can now use NIOTS.