Provide a channel handler using the client state machine (#601)
Motivation:

94e28d30 (#580) introduced a client state machine which enables us to
reduce the number of channel handlers the client has and confine much of
the stream state tracking to a single place. However, none of our
channel handlers were designed to use it!

Modifications:

Create a client channel handler using the aforementioned state machine
and update the state machine to talk in HTTP/2 concepts to avoid having
to use an additional handler to translate between HTTP/1 and HTTP/2
concepts.

Result:

Functionally, nothing, as the channel handler is not yet used.