Make 'GRPCStatus.Code' an enum (#942)
Motivation:

'GRPCStatus.Code' came straight from the old gRPC Swift implementation,
which in turn was copied almost verbatim from the grpc core library. It
includes the case 'doNotUse' which is documented as existing to force
users to include a default branch. Presumably, this is to ensure that
cases can be added in the future without breaking API. Unfortunately
that just doesn't cut it in Swift and we can enforce that a default
branch is provide by using a struct. This also allows us to add
additional status codes should any be added to the spec in the future.

Modifications:

- Turn `GRPCStatus.Code` into a `struct`, removing the `doNotUse` case
- Add a deprecated `doNotUse` case to the shims
- Add tests

Result:

- Programattically force a default branch when switching over
  'GRPCStatus.Code'