Use swift-argument-parser for examples etc (#1118)
Motivation:

We have a bunch of hand rolled argument parsing code for each example
and various tooling. We did this because we didn't want to burden users
of the 'GRPC' module with additional depdendencies.

Now that Swift 5.2 is our minimum supported version, we have target
based depdendency resolution. This means we can use depdendencies for
targets which aren't products of the package without burdening
dependents of the 'GRPC' module with unnecessary depdendencies.

Modifications:

- Use swift-argument-parser for all examples, performance and interop tests
- Update use of interop tests in CI

Result:

- Less hand rolled code
- Examples are easier to use: they have help messages and so on.