Write alloc counter test output to stdout and file (#1188)
Motivation:

The allocation counting test harness writes output to a file and then
parses the file to check for allocations. However, this step includes
building a whole bunch of stuff which can take some time and since no
output is generated during that time it appears to some CI systems as if
there is no forward progress. Travis will kill CI jobs if there is no
output for 10 minutes.

Modifications:

- 'tee' the output from running the allocation counting tests

Result:

We see progress on stdout.