Provide more context to failing RPCs from the connection pool (#1292)
Motivation:

When an RPC is started using the connection pool, it waits for a
connection to come up and a stream to be created. It may also time out
waiting for the connection to come up, this may be because the pool is
genuinely too busy to handle the RPC or there is an underlying problem
with the connection.

In the latter case it's not possible for the RPC to see the underlying
reason that it timed out. This makes issue diagnosis more difficult than
necessary when there root cause is at the connection level (e.g. due to TLS
configuration) as important but infrequent connection-level errors are
lost in the noise of RPC-level timeouts.

Modifications:

- Give connection pools a notion of a 'last known error', the most
  recent connection-level error to occur on a connection managed by the
  pool. The error is cleared as soon as any connection within the pool
  becomes available. This error is added to existing waiter errors to
  provide more context.
- Add a '_ConnectivityState' mirroring the public 'ConnectivityState'
  but which holds 'Error' as associated data in some cases.
- Wire this through the `ConnectionManager` to the `ConnectionPool`

Result:

Connection level errors are more obvious.