Remove client didSet state transition check (#1310)
Motivation:

`GRPCClientStateMachine.state` has a `didSet` to validate state
transitions. The presence of this block seems to trip up libbacktrace
when running with the scudo sanitizer. Moreover, the check is not
particularly helpful since we should always be transition to and from
the modifying state. This appears to have been leftover from when the
modifying state was introduced.

This is related to https://bugs.swift.org/browse/SR-15478.

Modifications:

Remove the didSet block for state in GRPCClientStateMachine.

Result:

We don't trip over in libbacktrace when running with scudo.