Deflake zero length write tests (#1317)
Motivation:

The zero-length write tests fail every so often because the XCTest
expectations are over fulfilled.

They can be overfulfilled when the client makes multiple attempts to
connect to the test server. This can happen because of happy eyeballs or
because the client is re-attempting to connect to the server.

Modifications:

- Use "127.0.0.1" instead of "localhost" to avoid using happy eyeballs
- Disable client connection re-establishment

Result:

Less flakey test.