[CodeGen Protobuf support] Message serializer and deserializer (#1764)
Motivation:

In the generated code we are specifying types for a serializer and deserializer. As we want to support SwiftProtobuf, we need to implement the generic serializer and deserializer for Protobuf messages, which will be the default if the user doesn't specify other IDL serializer and deserializer.

Modifications:

Created the serializer and deserializer structs conforming to `GRPCCore.MessageSerializer` and `GRPCCOre.MessageDeserializer` respectively and implemented their serialize and deserialize functions. Added tests for them.

Result:

The generated code can reference protobuf serializer and deserializer types, as they will be supported in GRPC.